--- const.go
@@ -1,21 +1,0 @@
-package main
-
-func getIgnoreExts() []string {
-	return []string{
-		".mkv",
-		".avi",
-		".mp4",
-		".mov",
-		".wmv",
-		".jpg",
-		".jpeg",
-		".png",
-		".gif",
-		".mp3",
-		".wav",
-		".pdf",
-		".heic",
-		".dmg",
-		".bin",
-	}
-}
